feat: Criando um arquivo de repository para cada uma das suas abstrações
